stochastic-planning-core-dev-cloudbuild/requirements.txt
fastapi
uvicorn
python-dotenv
mysql-connector-python
openai
google-generativeai
anthropic
groq
pydantic[email]
python-multipart
bcrypt
firebase-admin
python-jose[cryptography]
cryptography>=41.0.0
python-multipart
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-gcp-trace
opentelemetry-instrumentation
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-requests
redis
python-dateutil
google-cloud-firestore
google-auth
google-cloud-pubsub
google-cloud-logging